# Conflicts:
#	Platipus.Wallet.Api/StartupSettings/Filters/MockedErrorActionFilterAttribute.cs

+++ Platipus.Wallet.Api/Application/Requests/Test/DeleteErrorMockRequest.cs
@@ -7,10 +7,10 @@ using Infrastructure.Persistence;
 using Microsoft.EntityFrameworkCore;
 
 public record DeleteErrorMockRequest(
-    [property: DefaultValue("reevo_nazar_123")] string Username,
-    MockedErrorMethod Method) : IRequest<IPswResult>
+    [property: DefaultValue("reevo_platipus")] string Username,
+    MockedErrorMethod Method) : IRequest<IResult>
 {
-    public class Handler : IRequestHandler<DeleteErrorMockRequest, IPswResult>
+    public class Handler : IRequestHandler<DeleteErrorMockRequest, IResult>
     {
         private readonly WalletDbContext _context;
 
@@ -19,18 +19,25 @@ public record DeleteErrorMockRequest(
             _context = context;
         }
 
-        public async Task<IPswResult> Handle(
+        public async Task<IResult> Handle(
             DeleteErrorMockRequest request,
             CancellationToken cancellationToken)
         {
+            var userExists = await _context.Set<User>()
+                .Where(u => u.Username == request.Username)
+                .AnyAsync(cancellationToken);
+
+            if (!userExists)
+                return ResultFactory.Failure(ErrorCode.UserNotFound);
+
             var deletedErrorMocks = await _context.Set<MockedError>()
                 .Where(e => e.User.Username == request.Username && e.Method == request.Method)
                 .ExecuteDeleteAsync(cancellationToken: cancellationToken);
 
             if (deletedErrorMocks is 0)
-                return PswResultFactory.Failure(PswErrorCode.BadParametersInTheRequest);
+                return ResultFactory.Failure(ErrorCode.BadParametersInTheRequest);
 
-            return PswResultFactory.Success();
+            return ResultFactory.Success();
         }
     }
 }
